Create README.md
Leetlist is a program that was developed out of the need to quickly
modify an existing password list (whether generated through crush
or some other fasion) for a pen test. It allows the user to have a
base password list and modify it by capitalizing the first letter,
appending a prefix or suffix, and character substitution.